requirements.txt
Flask==0.10.1
RPi.GPIO==0.6.0a3
